+++ src/backend/base/langflow/api/utils.py
@@ -22,9 +22,7 @@ API_WORDS = ["api", "key", "token"]
 
 
 def has_api_terms(word: str):
-    return "api" in word and (
-        "key" in word or ("token" in word and "tokens" not in word)
-    )
+    return "api" in word and ("key" in word or ("token" in word and "tokens" not in word))
 
 
 def remove_api_keys(flow: dict):
@@ -34,11 +32,7 @@ def remove_api_keys(flow: dict):
             node_data = node.get("data").get("node")
             template = node_data.get("template")
             for value in template.values():
-                if (
-                    isinstance(value, dict)
-                    and has_api_terms(value["name"])
-                    and value.get("password")
-                ):
+                if isinstance(value, dict) and has_api_terms(value["name"]) and value.get("password"):
                     value["value"] = None
 
     return flow
@@ -59,9 +53,7 @@ def build_input_keys_response(langchain_object, artifacts):
             input_keys_response["input_keys"][key] = value
     # If the object has memory, that memory will have a memory_variables attribute
     # memory variables should be removed from the input keys
-    if hasattr(langchain_object, "memory") and hasattr(
-        langchain_object.memory, "memory_variables"
-    ):
+    if hasattr(langchain_object, "memory") and hasattr(langchain_object.memory, "memory_variables"):
         # Remove memory variables from input keys
         input_keys_response["input_keys"] = {
             key: value
@@ -71,97 +63,10 @@ def build_input_keys_response(langchain_object, artifacts):
         # Add memory variables to memory_keys
         input_keys_response["memory_keys"] = langchain_object.memory.memory_variables
 
-    if hasattr(langchain_object, "prompt") and hasattr(
-        langchain_object.prompt, "template"
-    ):
+    if hasattr(langchain_object, "prompt") and hasattr(langchain_object.prompt, "template"):
         input_keys_response["template"] = langchain_object.prompt.template
 
     return input_keys_response
-
-
-def update_frontend_node_with_template_values(frontend_node, raw_frontend_node):
-    """
-    Updates the given frontend node with values from the raw template data.
-
-    :param frontend_node: A dict representing a built frontend node.
-    :param raw_template_data: A dict representing raw template data.
-    :return: Updated frontend node.
-    """
-    if not is_valid_data(frontend_node, raw_frontend_node):
-        return frontend_node
-
-    # Check if the display_name is different than "CustomComponent"
-    # if so, update the display_name in the frontend_node
-    if raw_frontend_node["display_name"] != "CustomComponent":
-        frontend_node["display_name"] = raw_frontend_node["display_name"]
-
-    update_template_values(frontend_node["template"], raw_frontend_node["template"])
-
-    old_code = raw_frontend_node["template"]["code"]["value"]
-    new_code = frontend_node["template"]["code"]["value"]
-    frontend_node["edited"] = old_code != new_code
-
-    return frontend_node
-
-
-def raw_frontend_data_is_valid(raw_frontend_data):
-    """Check if the raw frontend data is valid for processing."""
-    return "template" in raw_frontend_data and "display_name" in raw_frontend_data
-
-
-def is_valid_data(frontend_node, raw_frontend_data):
-    """Check if the data is valid for processing."""
-
-    return (
-        frontend_node
-        and "template" in frontend_node
-        and raw_frontend_data_is_valid(raw_frontend_data)
-    )
-
-
-def update_template_values(frontend_template, raw_template):
-    """Updates the frontend template with values from the raw template."""
-    for key, value_dict in raw_template.items():
-        if key == "code" or not isinstance(value_dict, dict):
-            continue
-
-        update_template_field(frontend_template, key, value_dict)
-
-
-def update_template_field(frontend_template, key, value_dict):
-    """Updates a specific field in the frontend template."""
-    template_field = frontend_template.get(key)
-    if not template_field or template_field.get("type") != value_dict.get("type"):
-        return
-
-    if "value" in value_dict and value_dict["value"]:
-        template_field["value"] = value_dict["value"]
-
-    if "file_path" in value_dict and value_dict["file_path"]:
-        file_path_value = get_file_path_value(value_dict["file_path"])
-        if not file_path_value:
-            # If the file does not exist, remove the value from the template_field["value"]
-            template_field["value"] = ""
-        template_field["file_path"] = file_path_value
-
-
-def get_file_path_value(file_path):
-    """Get the file path value if the file exists, else return empty string."""
-    try:
-        path = Path(file_path)
-    except TypeError:
-        return ""
-
-    # Check for safety
-    # If the path is not in the cache dir, return empty string
-    # This is to prevent access to files outside the cache dir
-    # If the path is not a file, return empty string
-    if not str(path).startswith(user_cache_dir("langflow", "langflow")):
-        return ""
-
-    if not path.exists():
-        return ""
-    return file_path
 
 
 def validate_is_component(flows: list["Flow"]):
@@ -190,9 +95,7 @@ async def check_langflow_version(component: StoreComponentCreate):
 
     langflow_version = get_lf_version_from_pypi()
     if langflow_version is None:
-        raise HTTPException(
-            status_code=500, detail="Unable to verify the latest version of Langflow"
-        )
+        raise HTTPException(status_code=500, detail="Unable to verify the latest version of Langflow")
     elif langflow_version != component.last_tested_version:
         warnings.warn(
             f"Your version of Langflow ({component.last_tested_version}) is outdated. "
@@ -222,16 +125,12 @@ def format_elapsed_time(elapsed_time: float) -> str:
         return f"{minutes} {minutes_unit}, {seconds} {seconds_unit}"
 
 
-async def build_graph_from_db(
-    flow_id: str, session: Session, chat_service: "ChatService"
-):
+async def build_graph_from_db(flow_id: str, session: Session, chat_service: "ChatService"):
     """Build and cache the graph."""
     flow: Optional[Flow] = session.get(Flow, flow_id)
     if not flow or not flow.data:
         raise ValueError("Invalid flow ID")
-    graph = Graph.from_payload(
-        flow.data, flow_id, flow_name=flow.name, user_id=str(flow.user_id)
-    )
+    graph = Graph.from_payload(flow.data, flow_id, flow_name=flow.name, user_id=str(flow.user_id))
     for vertex_id in graph._has_session_id_vertices:
         vertex = graph.get_vertex(vertex_id)
         if vertex is None:
@@ -305,14 +204,10 @@ async def get_next_runnable_vertices(
     """
     async with chat_service._cache_locks[flow_id] as lock:
         graph.remove_from_predecessors(vertex_id)
-        direct_successors_ready = [
-            v for v in vertex.successors_ids if graph.is_vertex_runnable(v)
-        ]
+        direct_successors_ready = [v for v in vertex.successors_ids if graph.is_vertex_runnable(v)]
         if not direct_successors_ready:
             # No direct successors ready, look for runnable predecessors of successors
-            next_runnable_vertices = graph.find_runnable_predecessors_for_successors(
-                vertex_id
-            )
+            next_runnable_vertices = graph.find_runnable_predecessors_for_successors(vertex_id)
         else:
             next_runnable_vertices = direct_successors_ready
 
@@ -349,5 +244,4 @@ def parse_exception(exc):
     """Parse the exception message."""
     if hasattr(exc, "body"):
         return exc.body["message"]
-    return str(exc)
     return str(exc)
